bitbake: cache: Drop broken/unused code
Some parts of functions in Cache() were broken and unused, there was
also a totally unused function. This was historical as a result of the
cooker parsing process needing to handle cached entries in the main
thread but parsing actions in seperate processes.

Document the way it works, update the function name to be clear about
what it now does and drop the old code which was unused.

(Bitbake rev: af83ee32df85c8e4144f022a1f58493eb72cb18e)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>