bitbake: server: Ensure cooker profiling works
The previous cleanups meant that when the cooker was started, profiling
was always disabled as configuration was sent to the server later and this
was too late to profile the main loop.

Pass the "profile" option over the server commandline so that we can
profile cooker itself again, the setting can now take effect early enough.

(Bitbake rev: c1504f38f1a3619dfb59ab680cee27337a0e63b1)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>